Update batch tests for newer python versions.
Python 2.7.7 included a [fix](http://bugs.python.org/issue14983) to always add
a newline to multipart MIME messages. This makes one of our batch tests fail;
we could update to exactly require a newline, but there's no reason to be so
strict -- we might as well work across several python versions, since the code
in question is in python (not this library).