Support methodPath entries containing colon.
Currently, if we have a top-level methodPath of the form `path:methodName`,
we'll pass this to `urlparse.urljoin`, where this is interpreted as
`scheme:netloc`. In cases other than media upload, this is **not** what we
want.

This fixes the issue by adding our own custom `urljoin`, which detects and
handles this case. We also add a test.