Match travis matrix with tox testenv
Executing tox in the project directory is not the same of what is being
executed in travis-ci
This cause problems when replacing oauth2client library by google-auth
in the setup.py because oauth2client is mandatory for tests, and travis
was not installing oauth2client if it is not in the setup.py.
Also it is good to keep the consistency of the test environments.